safari-600.7-branch/LayoutTests/js/dictionary-no-cache-expected.txt

Test to ensure that we handle caching of prototype chains containing dictionaries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foundNewPrototypeProperty is true
PASS foundRemovedPrototypeProperty is false
PASS calledNewPrototypeSetter is true
PASS getTestProperty(test4) is "on prototype"
PASS getTestProperty(test4) is "on self"
PASS getTestProperty(test5) is "on prototype's prototype"
PASS getTestProperty(test5) is "on self"
PASS successfullyParsed is true

TEST COMPLETE
